move tests project to SSR, test clientOnly
