+++ packages/tests/app.config.ts
@@ -1,5 +1,3 @@
 import { defineConfig } from "@solidjs/start/config";
 
-export default defineConfig({
-  ssr: false
-});
+export default defineConfig();
