+++ packages/tests/src/routes/client-only/index.tsx
@@ -1,0 +1,12 @@
+import { clientOnly } from "@solidjs/start";
+
+const Component = clientOnly(() => import('./_component'))
+
+export default function App() {
+
+  return (
+    <>
+        <Component />
+    </>
+  );
+}
